Rename squashfs/squashfs to squashfs/low squashfs/low library name is now squashfslow
