+++ low/data/fullreader.go
