+++ low/inode/misc.go
