+++ low/inode/sym.go
